requirements.txt
asgiref==3.6.0
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.0
django-cors-headers==3.13.0
django-filter==22.1
django-rest-swagger==2.2.0
django-tenants==3.4.1
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.4
future==0.18.2
idna==3.4
importlib-metadata==5.2.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.1
MarkupSafe==2.1.1
openapi==1.1.0
openapi-codec==1.3.2
packaging==22.0
psycopg2==2.9.5
PyJWT==2.6.0
pyrsistent==0.19.2
pytz==2022.7
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
serializers==0.2.4
simplejson==3.18.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.13
zipp==3.11.0
